Created Config/ChipInfo structs and startEx
Created the StartEx function to initialize the chip to a custom
configuration.  Also created a type for the Config parameter to contain
the raw configuration strings sent to the device.  Created another
struct to hold the operation data for the driver (mainly the converted
parameters) and the socket table.  Started initial work on the
SocketOpen function to support opening of sockets. TODO: Add command
MACRO definitions.