Update API functions for Sockets/TCP
Updated the SocketOpen API to support an interface that allows the
re-sizing of Tx/Rx Buffer memory. the buffers remain fixed at 2 K each,
the parameters are for future expansion.  Renamed TcpStart to
TcpOpenClient to perform client connections in one function call rather
than several from the old (1.x) API. TcpStartServer and
TcpStartServerWait will be modified to TcpOpenServer and
TcpWaitForConnection.