Fixed run mode issues
There appears to be an issue with the W5500 at power up that might be
devout related, but was causing the initialization to fail. Modified
StartEX to initialize the W5500 in a loop with verify to ensure that
the initialization happens correctly.  Also fixed an issue with
ParseMAC that was causing the MAC address parsing to fail incorrectly.
This was related to incrementing the index within a MACRO function.
Changed from dash separators to colons for MAC addressing.